requirements.txt
psutil
pyrogram==2.0.106
requests==2.31.0
tgcrypto==1.2.5
asyncio==3.4.3
python-dotenv==1.0.0
motor
pyromod
aiohttp
apscheduler
pillow


click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
